+++ bitmath/integrations/argparse.py
@@ -26,10 +26,8 @@
 
 import bitmath
 import argparse
-import progressbar.widgets
 
-######################################################################
-# Integrations with 3rd party modules
+
 def BitmathType(bmstring):
     """An 'argument type' for integrations with the argparse module.
 
@@ -78,27 +76,3 @@ examples to conserve single quotes in the parse_args call):
                                          bmstring)
     else:
         return argvalue
-
-######################################################################
-# Speed widget for integration with the Progress bar module
-class BitmathFileTransferSpeed(progressbar.widgets.Widget):
-    """Widget for showing the transfer speed (useful for file transfers)."""
-    __slots__ = ('system', 'format')
-
-    def __init__(self, system=bitmath.NIST, format="{value:.2f} {unit}/s"):
-        self.system = system
-        self.format = format
-
-    def update(self, pbar):
-        """Updates the widget with the current NIST/SI speed.
-
-Basically, this calculates the average rate of update and figures out
-how to make a "pretty" prefix unit"""
-
-        if pbar.seconds_elapsed < 2e-6 or pbar.currval < 2e-6:
-            scaled = bitmath.Byte()
-        else:
-            speed = pbar.currval / pbar.seconds_elapsed
-            scaled = bitmath.Byte(speed).best_prefix(system=self.system)
-
-        return scaled.format(self.format)
